refactor: align settings presentation package with its directory
Keep Settings Compose screens in settings.presentation. Shared
scaffolds stay in ui.compose.
