+++ app/src/main/java/com/dpis/module/ui/presentation/wear/WearSecondaryPages.kt
@@ -21,6 +21,7 @@ import androidx.wear.compose.material3.Text
 import com.dpis.module.R
 import com.dpis.module.about.OpenSourceLicenseItem
 import com.dpis.module.about.presentation.LicenseDetailDialog
+import com.dpis.module.settings.presentation.InterfaceScaleDialogContent
 import com.dpis.module.settings.AppUiScaleManager
 import com.dpis.module.settings.ThemeModeStore
 import com.dpis.module.ui.presentation.WearMaterialTheme
